+++ src/language/helpers/shortcuts.ts
