+++ src/language/validation/other/imports.ts
@@ -1,6 +1,6 @@
 import { ValidationAcceptor } from 'langium';
 import { SdsImportAlias } from '../../generated/ast.js';
-import { isWildcardImport } from '../../ast/checks.js';
+import { isWildcardImport } from '../../helpers/checks.js';
 
 export const CODE_IMPORT_WILDCARD_IMPORT_WITH_ALIAS = 'import/wildcard-import-with-alias';
