+++ src/language/validation/other/types/unionTypes.ts
@@ -1,6 +1,6 @@
 import { SdsUnionType } from '../../../generated/ast.js';
 import { ValidationAcceptor } from 'langium';
-import { typeArgumentsOrEmpty } from '../../../ast/shortcuts.js';
+import { typeArgumentsOrEmpty } from '../../../helpers/shortcuts.js';
 import { isEmpty } from 'radash';
 
 export const CODE_UNION_TYPE_MISSING_TYPE_ARGUMENTS = 'union-type/missing-type-arguments';
